일부 로직 구현

+++ src/main/java/kr/ac/deu/computer_engineering/Absenteeism/Management/route/rank/RankController.java
@@ -9,11 +9,16 @@ import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.Rank.dto.Ran
 import kr.ac.deu.computer_engineering.Absenteeism.Management.domain.Rank.dto.RankDto;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.service.rank.RankService;
 import kr.ac.deu.computer_engineering.Absenteeism.Management.service.team.TeamService;
+import kr.ac.deu.computer_engineering.Absenteeism.Management.utils.RoleValidate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -25,42 +30,70 @@ public class RankController {
     @Tag(name = "직급")
     @Operation(summary = "직급 목록 조회", description = "직급 목록 조회")
     @GetMapping("")
-    public ResponseEntity<?> getItemList(@RequestParam(required = false) String q) {
-        List<Rank> result = rankService.getList(q);
-        return new ResponseEntity<>(result, HttpStatus.OK);
+    public ResponseEntity<?> getItemList(HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            List<Rank> result = rankService.getList(session);
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @Tag(name = "직급")
     @Operation(summary = "직급 Id로 상세 조회", description = "직급 상세 조회")
     @GetMapping("/{rankId}")
-    public ResponseEntity<?> getItemById(@PathVariable Long rankId) {
-        Rank rank = rankService.getRankById(rankId);
-        return new ResponseEntity<>(rank, HttpStatus.OK);
+    public ResponseEntity<?> getItemById(
+            HttpServletRequest request,
+            @PathVariable Long rankId) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            Rank rank = rankService.getRankById(rankId);
+            return new ResponseEntity<>(rank, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @Tag(name = "직급")
     @Operation(summary = "직급 정보 생성", description = "직급 정보 생성")
     @PostMapping("")
-    public ResponseEntity<?> createItem(@RequestBody RankDto dto) {
-        rankService.createRank(dto);
-        return new ResponseEntity<>(HttpStatus.CREATED);
+    public ResponseEntity<?> createItem(
+            HttpServletRequest request,
+            @Valid @RequestBody RankDto dto,
+            BindingResult bindingResult) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            rankService.createRank(dto);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @Tag(name = "직급")
     @Operation(summary = "직급 Id로 직급 정보 수정", description = "직급 정보 수정")
     @PutMapping("/{rankId}")
     public ResponseEntity<?> updateItemById(
+            HttpServletRequest request,
             @PathVariable Long rankId,
-            @RequestBody RankDto dto) {
-        rankService.updateRank(rankId, dto);
-        return new ResponseEntity<>(HttpStatus.OK);
+            @Valid @RequestBody RankDto dto) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            rankService.updateRank(rankId, dto);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @Tag(name = "직급")
     @Operation(summary = "직급 Id로 직급 정보 삭제", description = "직급 정보 삭제")
     @DeleteMapping("/{rankId}")
-    public ResponseEntity<?> deleteItemById(@PathVariable Long rankId) {
-        rankService.deleteRank(rankId);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<?> deleteItemById(
+            HttpServletRequest request,
+            @PathVariable Long rankId) {
+        HttpSession session = request.getSession();
+        if (RoleValidate.isRoleCeo(session) || RoleValidate.isRoleManager(session)) {
+            rankService.deleteRank(rankId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
